// backend-game-runner/src/infrastructure/grpc/proto/types/playfulbot/v0/FollowGameResponse.ts
// Original file: src/grpc/proto/playfulbot/v0/playfulbot_v0.proto

import type { Game as _playfulbot_v0_Game, Game__Output as _playfulbot_v0_Game__Output } from './Game';
import type { GamePatch as _playfulbot_v0_GamePatch, GamePatch__Output as _playfulbot_v0_GamePatch__Output } from './GamePatch';
import type { GameCanceled as _playfulbot_v0_GameCanceled, GameCanceled__Output as _playfulbot_v0_GameCanceled__Output } from './GameCanceled';

export interface FollowGameResponse {
  'game'?: (_playfulbot_v0_Game);
  'patch'?: (_playfulbot_v0_GamePatch);
  'canceled'?: (_playfulbot_v0_GameCanceled);
  'gameOrUpdate'?: "game"|"patch"|"canceled";
}

export interface FollowGameResponse__Output {
  'game'?: (_playfulbot_v0_Game__Output);
  'patch'?: (_playfulbot_v0_GamePatch__Output);
  'canceled'?: (_playfulbot_v0_GameCanceled__Output);
  'gameOrUpdate': "game"|"patch"|"canceled";
}